Fix: Add description to _memset

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -3,12 +3,13 @@
  * @s: Pointer to the memory area
  * @n: Num of bytes to be filled
  * @b: The constant byte
+ * Return: Pointer to the mem
  */
 char *_memset(char *s, char b, unsigned int n)
 {
-	int i;
+	unsigned int i;
 
-	for (i = 0; i < (int)n; i++)
+	for (i = 0; i < n; i++)
 	{
 		s[i] = b;
 	}
